feat: support float minutes for autoclose timeouts
Allows sub-minute values (e.g. 0.25 = 15s) via AUTOCLOSE_DONE_MINUTES
and AUTOCLOSE_DEAD_MINUTES env vars.

+++ src/ccgram/config.py
@@ -254,10 +254,10 @@ class Config:
         self.send_max_results: int = _parse_int_env("CCGRAM_SEND_MAX_RESULTS", 50)
 
     def _init_lifecycle(self) -> None:
-        self.autoclose_done_minutes: int = int(
+        self.autoclose_done_minutes: float = float(
             os.getenv("AUTOCLOSE_DONE_MINUTES", "30")
         )
-        self.autoclose_dead_minutes: int = int(
+        self.autoclose_dead_minutes: float = float(
             os.getenv("AUTOCLOSE_DEAD_MINUTES", "10")
         )
         self.pane_lifecycle_notify: bool = os.getenv(
